Add files via upload
This script detects the Rossmann logo, and then writes to the command.txt file for the wifi-car.py script to pick up and control the vehicle.  

if logo is not detected the vehicle stops.  code works but is janky.  there is a lag between logo detection and commands being triggered.

next version will skip wifi-car.py ant command file to directly trigger gpis pins.